implement platform-specific IPC mechanisms for Windows and Unix-like systems
